Convert barline output data to mei
given the user input of staff groups, transform the output data of
gabriel's barfinding algorithm: <staffnum ulx uly lrx lry> to mei.
Currently the program does not encode system breaks or situations where
multiple movements are on the page or instruments are subtracted or
added on the page, but its a good enough start for the meeting tomorrow.