Handle hidden staves
Parse staff group structures where there are hidden staves within a
system. For example, a system of piano followed by a new system of piano
with vocals overtop---in the first system the vocal staves are hidden.
The encoded staff grouping in the mei file will take the system
structure with the largest number of staves.